Set the validator for formatters at plugin load, not in autoload.
Previously, if any formatters were registered before
autoload/codefmt.vim was sourced, the call to set the validator would
attempt to load the autoload file, ending in infinite recursion.

This wasn't previously detected because the call in the Vroom tests to
e.g. codefmt#SetWhetherToPerformIsAvailableChecksForTesting() caused the
autoload/ file to be sourced first.  Splitting the tests out into a
separate file (and removing that now-unnecessary call) means that the
tests now cover this.

Also reorganised the tests to be slightly clearer and removed the test
that showed that AddExtension() can throw WrongType, since that's
already covered by tests of AddExtension() itself.

(We're still not testing the case where we load Maktaba, register an
extension, and then load codefmt. That's hard to do given the way that
we use bootstrap.vim - which loads both - in the tests.)