feat(prettier): use npx for project-local prettier
If npx is available, use it in preference to calling prettier directly.
This enables project-local prettier, which is important for teams to use
matching versions.

This doesn't change anything for non-project prettier installs, because
npx will fall back to global prettier.

Since the existence of npx doesn't actually imply the existence of
prettier, we have to call it once to find out if it's usable. Since this
is expensive (compared to not calling it, anyway), cache the result as
s:prettier_is_available and invalidate if prettier_executable changes.