clang-format: work around vim/vim#5930 when positioning cursor
This bug (and a related predecessor) cause `:goto` and `line2bytes()`
to do the wrong thing when vim textprops are used.
This manifests as "cursor jumps around randomly after formatting" when using
plugins that use textprops for diagnostics, highlighting etc.

We happen to have all the code in an array when we need to do cursor coordinate
conversions, so we do those in vimscript instead of querying vim.

Without this patch, The newly added test fails on affected versions of vim
including both vim master and debian's 8.1.2269 (in different ways).

Vim bug tracker entry:
https://github.com/vim/vim/issues/5930
Similar workaround in first-party vim integration:
https://github.com/llvm/llvm-project/commit/591be7ec500c151d9232366042e21c74e006292c