Fix dartfmt's IsAvailable() method.
d927656 changed the dartfmt_executable flag from a string to a list, but
didn't change the use inside IsAvailable(), leading to:

Failed to evaluate whether formatter dartfmt is available: Vim(return)
:E730: Using a List as a String

The change uses the ResolveFlagToArray() helper in both IsAvailable()
and FormatRange() (which incidentally also allows defining the flag
using a Function), and updates the documentation for all flags using
ResolveFlagToArray() to use the same phrasing.

Fixes #190.